Just a thought.
You can get rid of this pull request if you want, I figured since "glod" was a funny mispronunciation we could do the same for the other ores, but nothing funny comes to mind for silver so I don't think it's necessary, especially since we're only replacing the things we want to change. I'll finish editing the lang file and copy all the stuff I decided to change in a later request.