Lang edits
Took all the things from the Lang file I thought were funny. The things like "Diemond" and "Iorn" are just changes based on Glod that sound funny, but also trying to keep in the same style. Also because I always die when trying to mine diamond. Other stuff is just stuff I found funny, but a lot are jokes from LRRchat and stuff. The "droppa" and "hoppa" are propa Orky, of course.